Add fixtures to pycompanies/tests/fixtures.py
