+++ pycompanies/tests/test_views.py
@@ -1,10 +1,11 @@
 import pytest
+
 from django.contrib.messages import get_messages as contrib_get_messages
-from django.test import Client
 from django.urls import reverse_lazy
 
-from events.tests.factories import UserFactory, DEFAULT_USER_PASSWORD
-from .factories import CompanyFactory, UserCompanyProfileFactory
+from pycompanies.tests.factories import CompanyFactory, UserCompanyProfileFactory, UserFactory
+from pycompanies.tests.fixtures import (create_client, create_logged_client, create_user, # noqa
+                                        create_user_company_profile)
 
 
 ERROR_USER_DOES_NOT_EXIST = 'Le usuarie que ingresó no existe.'
@@ -19,31 +20,6 @@ def get_plain_messages(request):
     """
     messages = contrib_get_messages(request.wsgi_request)
     return [m.message for m in messages]
-
-
-@pytest.fixture(name='client')
-def create_client():
-    return Client()
-
-
-@pytest.fixture(name='user')
-def create_user():
-    return UserFactory()
-
-
-@pytest.fixture(name='logged_client')
-def create_logged_client(user):
-    client = Client()
-    client.login(username=user.username, password=DEFAULT_USER_PASSWORD)
-    return client
-
-
-@pytest.fixture(name='user_company_profile')
-def create_user_company_profile():
-    """
-    Fixture with a dummy UserCompanyProfileFactory
-    """
-    return UserCompanyProfileFactory.create()
 
 
 @pytest.mark.django_db
@@ -63,12 +39,11 @@ def test_associate_nonexistent_user(logged_client):
 
 
 @pytest.mark.django_db
-def test_associate_user_in_company(logged_client):
+def test_associate_user_in_company(logged_client, user):
     """
     Should redirect and send a success message after associating the user
     """
     company = CompanyFactory.create()
-    user = UserFactory.create()
 
     ASSOCIATE_URL = reverse_lazy('companies:associate', kwargs={'company': company.id})
 
@@ -81,12 +56,11 @@ def test_associate_user_in_company(logged_client):
 
 
 @pytest.mark.django_db
-def test_associate_user_already_in_company(logged_client):
+def test_associate_user_already_in_company(logged_client, user):
     """
     Should fail, redirect and send a warning message if the user is already in the company
     """
     company = CompanyFactory.create()
-    user = UserFactory.create()
     user_company = UserCompanyProfileFactory.create(company=company, user=user)
 
     ERROR_USER_ALREADY_IN_COMPANY = ('Le usuarie que desea vincular ya '
@@ -103,11 +77,10 @@ def test_associate_user_already_in_company(logged_client):
 
 
 @pytest.mark.django_db
-def test_associate_user_in_other_company(logged_client):
+def test_associate_user_in_other_company(logged_client, user):
     """
     Should fail, redirect and send a warning message if the user is associated in other company
     """
-    user = UserFactory.create()
     company = CompanyFactory.create(name='company_1')
     other_company = CompanyFactory.create(name='company_2')
     UserCompanyProfileFactory.create(company=other_company, user=user)
@@ -193,8 +166,8 @@ def test_company_disassociate_last_user_from_company(logged_client, user):
     """
     Message in context should notice if is the last user associated to the company
     """
-    DISASSOCIATE_MESSAGE = "Esta es la última persona vinculada a esta empresa "\
-        "¿Estás seguro que deseas desvincularla?"
+    DISASSOCIATE_MESSAGE = ('Esta es la última persona vinculada a esta empresa '
+                            '¿Estás seguro que deseas desvincularla?')
 
     company_1 = CompanyFactory.create(name='company_1')
     user_company_profile = UserCompanyProfileFactory.create(company=company_1, user=user)
@@ -218,7 +191,7 @@ def test_company_disassociate_one_user_from_company(logged_client, user):
     user_company_profile = UserCompanyProfileFactory.create(company=company, user=user)
     UserCompanyProfileFactory.create(company=company, user=user_2)
 
-    DISASSOCIATE_MESSAGE = f"¿Estás seguro que desea desvincular a {user} de {company.name}?"
+    DISASSOCIATE_MESSAGE = f'¿Estás seguro que deseas desvincular a {user} de {company.name}?'
 
     COMPANY_DISSASOCIATE_URL = reverse_lazy('companies:disassociate',
                                             kwargs={'pk': user_company_profile.id})
